src/apps/everything2md/requirements.txt
mcp[cli]
python-dotenv
pymupdf4llm
pptx2md
fastapi
uvicorn
python-multipart
